Create an initial migration so db:migrate works
When NUCore was brought in to git, there was a pre-existing schema
which was not captured in the migrations, thus `rake db:migrate` would
not work from a blank database.

I tried two techniques in order to get `rake db:migrate` working and
decided on this one.

The first:
* Find the first commit’s schema.rb and put that in an initial migration.
* Delete all migrations up to that point
* Fix any issues caused by missing data

That was fairly simple, but it lost much of the history of the project,
so I tried another process.

* Find the first commit's schema.rb and put that in an intial migration.
* Run `rake db:drop db:create db:migrate` and as I found errors, correct
the original migration accordingly.
* Fix issues caused by missing data or API changes in the migrations

They are similar, but I'm happier with this result.

* It gives a better history of the project
* Index names are more in line with Rails standards (there are still a
few remaining `sys_xxx` coming from the original oracle installation to
be cleaned up)

Caveats:

* I still need to test against Oracle.
* Merging this in to the current deployments will likely cause issues
when they run `db:migrate` since it will try to run the new migration.

Notes on the changes:
* Some columns have changed ordering within the tables -- not an issue
for new deploys, but could result in weirdness on existing development
commits when working off of existing databases
* Timestamps for some tables now have `null: false`. I think a separate
migration can bring this to a consistent state.
* The new version of foreigner uses 1.9 hash syntax, so `add_foreign_key`
in `schema.rb` will be using the new syntax while the rest of schema.rb
uses the `=>` syntax.

+++ db/migrate/20110426221701_remove_reviewable_state.rb
@@ -1,17 +1,23 @@
 class RemoveReviewableState < ActiveRecord::Migration
 
   def self.up
-    reviewable, inprocess=OrderStatus.find_by_name('Reviewable'), OrderStatus.inprocess.first
-    execute("UPDATE order_details SET order_status_id=#{inprocess.id},state='inprocess' WHERE order_status_id=#{reviewable.id} AND state='reviewable'")
+    reviewable = OrderStatus.find_by_name('Reviewable')
+    inprocess = OrderStatus.inprocess.first
+    if inprocess
+      execute("UPDATE order_details SET order_status_id=#{inprocess.id},state='inprocess' WHERE order_status_id=#{reviewable.id} AND state='reviewable'")
+    end
 
-    right, left=reviewable.right, reviewable.left
-    diff=right-left
+    if reviewable
+      right reviewable.right
+      left = reviewable.left
+      diff = right-left
 
-    # awesome_nested_set barfs trying to fix descendants on Oracle, so skip its shoddy work and rebuild tree manually
-    reviewable.skip_before_destroy=true
-    reviewable.destroy
+      # awesome_nested_set barfs trying to fix descendants on Oracle, so skip its shoddy work and rebuild tree manually
+      reviewable.skip_before_destroy = true
+      reviewable.destroy
 
-    execute("UPDATE order_statuses SET lft=(lft-(#{diff}+1)), rgt=(rgt-(#{diff}+1)) WHERE lft > #{right}")
+      execute("UPDATE order_statuses SET lft=(lft-(#{diff}+1)), rgt=(rgt-(#{diff}+1)) WHERE lft > #{right}")
+    end
   end
 
   def self.down
